+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # @project-version@ (@build-time@)
 
+* Added support for WoW Forever.
 * Slightly reduced the default font size of the energy bar.
-
+* Changed the addon compartment entry to only show up on Druids, which also fixes a Lua error when clicking it on other classes.
+* Added a chat message when using /bitw on a non-Druid character.